feat(ios/engine): unit tests for link parsing, search closure calling
